codal_port/microbit_pinaudio: Only allow one pin for audio selection.
The audio, music and speech modules can now only take either None ore a
single pin object to specify the output for the audio.  Note that this is
in addition to the speaker output which is now a completely independent
audio entity, and to which audio will always be directed (speaker output
can be disabled via speaker.off()).

Signed-off-by: Damien George <damien@micropython.org>