codal_port: Stop code execution if a run_every raises an exception.
If a callback scheduled by microbit.run_every() raises an uncaught
exception, then that run_every will be cancelled.  If this happens while
the initial main.py is executing it will propagate a SystemExit which will
stop the main program and drop to the REPL.  The original exception from
the run_every will then be scrolled on the display in the same way as
exceptions from main.py are.

When the main program finishes (either by completion or because it raised
an uncaught exception) then all run_every callbacks will be cancelled.

Addresses issues #90 and #94.

Signed-off-by: Damien George <damien@micropython.org>