codal_port/modradio: Turn the radio on when importing radio module.
In all cases radio.on() is necessary to start using the radio, otherwise
functions like radio.send() or radio.receive() will throw an exception
"ValueError: radio is not enabled".  The majority of the micro:bit
features, especially those that can be turned off like the display, are
"on" by default.  So make the radio consistent in that regard.

Addresses issue #97.

Signed-off-by: Damien George <damien@micropython.org>