codal_app: Set 8bit level detector floor to 52.0 dB.
To match the same floor as the previous implementatio. As the new
implementation is more sensitive to lower volume levels and in
"generic quiet room" it gave much higher values, increasing
the LEVEL_DETECTOR_SPL_8BIT_000_POINT config value will match
the previous version.

As discussed in
https://github.com/lancaster-university/codal-microbit-v2/issues/236

Can be removed with the next CODAL tag 0.2.44, which will change
the default value of LEVEL_DETECTOR_SPL_8BIT_000_POINT.